refactor: 와일드 카드 제거
