+++ domain/src/main/kotlin/Application.kt
@@ -1,6 +1,4 @@
 import controller.GameController
-import domain.RenjuRule
-import domain.RenjuRuleAdapter
 
 fun main() {
     Thread.setDefaultUncaughtExceptionHandler { _, exception ->
