+++ domain/src/main/kotlin/controller/PointMapper.kt
@@ -3,7 +3,7 @@ package controller
 import domain.Point
 import dto.PointDTO
 
-object PointMapper : Mapper<Point,PointDTO> {
+object PointMapper : Mapper<Point, PointDTO> {
     override fun domainToDTO(source: Point): PointDTO {
         return PointDTO(source.x, source.y)
     }
