+++ domain/src/main/kotlin/domain/Coordinate.kt
@@ -12,10 +12,12 @@ data class Coordinate private constructor(val point: Point) {
 
     companion object {
         fun from(x: Int, y: Int): Coordinate {
-            require(x < Board.BOARD_END_POINT.x ||
+            require(
+                x < Board.BOARD_END_POINT.x ||
                     y < Board.BOARD_END_POINT.y ||
                     x >= Board.BOARD_START_POINT.x ||
-                    y >= Board.BOARD_START_POINT.y) { ERROR_NOT_FOUNT_COORDINATE }
+                    y >= Board.BOARD_START_POINT.y
+            ) { ERROR_NOT_FOUNT_COORDINATE }
             return Coordinate(Point(x, y))
         }
