+++ domain/src/main/kotlin/domain/PlaceResult.kt
