+++ domain/src/main/kotlin/domain/Point.kt
@@ -1,7 +1,7 @@
 package domain
 
-data class Point( val x: Int,  val y: Int) {
-     operator fun plus(other: Point): Point {
+data class Point(val x: Int, val y: Int) {
+    operator fun plus(other: Point): Point {
         return Point(x + other.x, y + other.y)
     }
