+++ domain/src/main/kotlin/domain/RenjuRuleInterface.kt
@@ -3,5 +3,5 @@ package domain
 interface RenjuRuleInterface {
     fun isThreeToThree(stone: Stone, stones: Stones): Boolean
     fun isFourToFour(stone: Stone, stones: Stones): Boolean
-    fun isOverFive(stoneScore : Int) : Boolean
+    fun isOverFive(stoneScore: Int): Boolean
 }
