+++ domain/src/test/kotlin/domain/BoardTest.kt
@@ -3,7 +3,6 @@ package domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
-import org.junit.jupiter.api.assertThrows
 
 class BoardTest {
     @Test
